move mask to a separate file
